+++ src/geo/temporal_spatial_query.cpp
@@ -158,7 +158,7 @@ TemporalSpatialQuery::entitiesWithinDistanceAtTime(
         std::vector<std::pair<std::string, GeometryInfo>> geo_entries;
         geo_entries.reserve(rows.size());
 
-        for (std::size_t i = 0; i <static_cast<int>(rows.size()); ++i) {
+        for (std::size_t i = 0; i < rows.size(); ++i) {
             auto geom = extractGeometry(rows[i], geo_field);
             if (!geom.has_value()) {
                 continue;
